Add a ChordN gate extension mode
ChordN mode provides an interesting way to turn your notes into a chord
progression. These new features provide a mechanism to extend the gate
beyond the current chord for notes which appear in the next chord too.
This allows for a (potentially) more fluid and compelling playing of the
notes and chords.

The two parameters introduced allow the user to choose how many chords a
note may be extended through, and the maximum number of time divisions
any extension can last. The defaults are conservative, but enough to
give a feel for what you can get from the mechanism.